fix: v1.12.2 — API deserialization error for extra_usage credits fields
long? → double? for used_credits and monthly_limit in ExtraUsage model
to handle floating-point values returned by the Anthropic API.

+++ ClaudeUsageTray/Models/UsageData.cs
@@ -44,10 +44,10 @@ public class ExtraUsage
     public double? Utilization { get; set; }
 
     [JsonPropertyName("used_credits")]
-    public long? UsedCredits { get; set; }
+    public double? UsedCredits { get; set; }
 
     [JsonPropertyName("monthly_limit")]
-    public long? MonthlyLimit { get; set; }
+    public double? MonthlyLimit { get; set; }
 }
 
 public class SessionStats
